LPS-56785 Make portlet names globally unique
Otherwise, resource assignment won't work correctly. Also, create two
portlet classes so that portlet names refer to valid classes (following
OSGi naming conventions).
