+++ modules/apps/microblogs/microblogs-web/docroot/WEB-INF/portlet.xml
@@ -2,7 +2,7 @@
 
 <portlet-app xmlns="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd http://java.sun.com/xml/ns/portlet/portlet-app_2_0.xsd" version="2.0">
 	<portlet>
-		<portlet-name>1</portlet-name>
+		<portlet-name>com_liferay_microblogs_web_microblogs_portlet_MicroblogsPortlet</portlet-name>
 		<display-name>Microblogs</display-name>
 		<portlet-class>com.liferay.microblogs.web.microblogs.portlet.MicroblogsPortlet</portlet-class>
 		<init-param>
@@ -33,9 +33,9 @@
 		</security-role-ref>
 	</portlet>
 	<portlet>
-		<portlet-name>2</portlet-name>
+		<portlet-name>com_liferay_microblogs_web_microblogs_portlet_MicroblogsStatusUpdatePortlet</portlet-name>
 		<display-name>Microblogs Status Update</display-name>
-		<portlet-class>com.liferay.microblogs.web.microblogs.portlet.MicroblogsPortlet</portlet-class>
+		<portlet-class>com.liferay.microblogs.web.microblogs.portlet.MicroblogsStatusUpdatePortlet</portlet-class>
 		<init-param>
 			<name>config-jsp</name>
 			<value>/html/status_update/configuration.jsp</value>
