+++ modules/apps/microblogs/microblogs-web/docroot/WEB-INF/src/com/liferay/microblogs/web/microblogs/portlet/MicroblogsStatusUpdatePortlet.java
@@ -12,19 +12,10 @@
  * details.
  */
 
-package com.liferay.microblogs.constants;
+package com.liferay.microblogs.web.microblogs.portlet;
 
 /**
- * @author Jonathan Lee
- * @author Joan Kim
+ * @author Adolfo Pérez
  */
-public class MicroblogsPortletKeys {
-
-	public static final String MICROBLOGS =
-		"com_liferay_microblogs_web_microblogs_portlet_MicroblogsPortlet";
-
-	public static final String MICROBLOGS_STATUS_UPDATE =
-		"com_liferay_microblogs_web_microblogs_portlet_" +
-			"MicroblogsStatusUpdatePortlet";
-
+public class MicroblogsStatusUpdatePortlet extends BaseMicroblogsPortlet {
 }
